fix: renamed data recording to transaction entry and separated the transaction entry frontend files
