+++ frontend/lib/main.dart
@@ -8,7 +8,7 @@ import 'models/ui/register.dart';
 import 'models/ui/home.dart';
 import 'models/ui/view_profile.dart';
 import 'models/ui/edit_profile.dart';
-import 'models/ui/data_recording.dart';
+import 'models/ui/transaction_entry.dart';
 import 'models/ui/inventory_page.dart';
 import 'models/ui/item_details.dart';
 
@@ -32,7 +32,7 @@ class MyApp extends StatelessWidget{
         '/view_profile': (context) => ViewProfilePage(),
         '/edit_profile': (context) => EditProfilePage(),
         '/dashboard': (context) => Dashboard(),
-        '/data_recording': (context) => DataRecording(),
+        '/transaction_entry': (context) => TransactionEntryPage(),
         '/inventory': (context) => InventoryPage(),
         '/stock_alert': (context) => StockAlert()
