+++ frontend/pubspec.yaml
@@ -30,13 +30,13 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  csv: ^5.0.0
+  csv: ^6.0.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   fl_chart: ^1.0.0
   http: ^1.4.0
-  intl: ^0.19.0
+  intl: ^0.20.2
   flutter_dotenv: ^5.1.0 
   flutter_secure_storage: ^9.2.4
